move components to domain/components
