+++ compass_app/app/lib/ui/auth/logout/view_models/logout_viewmodel.dart
@@ -1,4 +1,4 @@
-import '../../../../data/components/auth/auth_logout_component.dart';
+import '../../../../domain/components/auth/auth_logout_component.dart';
 import '../../../../utils/command.dart';
 import '../../../../utils/result.dart';
