+++ compass_app/app/test/domain/components/auth/auth_logout_component_test.dart
@@ -1,4 +1,4 @@
-import 'package:compass_app/data/components/auth/auth_logout_component.dart';
+import 'package:compass_app/domain/components/auth/auth_logout_component.dart';
 import 'package:compass_model/model.dart';
 import 'package:flutter_test/flutter_test.dart';
