+++ compass_app/app/test/ui/booking/booking_screen_test.dart
@@ -1,5 +1,5 @@
-import 'package:compass_app/ui/booking/components/booking_create_component.dart';
-import 'package:compass_app/ui/booking/components/booking_share_component.dart';
+import 'package:compass_app/domain/components/booking/booking_create_component.dart';
+import 'package:compass_app/domain/components/booking/booking_share_component.dart';
 import 'package:compass_app/ui/booking/view_models/booking_viewmodel.dart';
 import 'package:compass_app/ui/booking/widgets/booking_screen.dart';
 import 'package:compass_model/model.dart';
